# perf: Optimize the command encoding