Not as important as `IO` latency but can still be a decent throughput improvement.

(Lots of variance in the benchmarks but there is still some clear improvements)
```
encode/pipeline         time:   [727.07 us 731.89 us 737.21 us]
                        change: [-59.635% -58.716% -57.668%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) high mild
  4 (4.00%) high severe
encode/pipeline_nested  time:   [246.61 us 249.05 us 251.81 us]
                        change: [-71.597% -70.950% -70.333%] (p = 0.00 < 0.05)
                        Performance has improved.
encode/integer          time:   [693.79 us 698.95 us 703.98 us]
                        change: [-63.633% -63.125% -62.651%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild
encode/small            time:   [1.1458 us 1.1580 us 1.1701 us]
                        change: [-40.826% -39.741% -38.664%] (p = 0.00 < 0.05)
                        Performance has improved.
```

I could probably refactor to remove these breaking changes but for the moment...

BREAKING CHANGE

`ToRedisArgs` has been changed to take take an instance of `RedisWrite`  instead of `Vec<Vec<u8>>`. Use the `write_arg` method instead of `Vec::push`.

`impl Iterator` is used which requires 1.26

`iter` now takes `self` by value instead of cloning `self` inside the method.