@Marwes 2018-09-12
If #167 is merged a breaking release is needed anyways so perhaps it is fine to use `impl Iterator` then as well.