@Marwes 2019-02-18
I rebased and squashed this since there was some files accidentally included and to fix the conflicts.