@badboy 2019-02-18
Test failures due to formatting (because I merged #167). I'm fine with merging as is and fixing afterwards.