@badboy 2019-03-16
Ah damn, of course this has merge conflicts now and thus will need a rebase (and resolving those conflicts).

But otherwise I would be fine with merging it now.